% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexpressionNetwork.r
\name{CoexpressionNetwork_BuildNetwork}
\alias{CoexpressionNetwork_BuildNetwork}
\title{Building the coexpression network}
\usage{
CoexpressionNetwork_BuildNetwork(
  networkType = "unsigned",
  power_thre,
  maxBlockSize,
  cutHeight,
  workingDir,
  datasetDir,
  cores
)
}
\arguments{
\item{networkType}{Type of Network to be created. Default = "unsinged".}

\item{power_thre}{Power Threshold for soft thresholding during network construction. The current version of NetRank does not include functionality to help the user choose the power threshold. For more information see the WGCNA publication.}

\item{maxBlockSize}{Maximum block size. Higher values demand more RAM.}

\item{cutHeight}{Cut height for culling outliers. Choose value by checking plots created by CoexpressionNetwork_Preprocessing. For further information see the WGCNA publication}

\item{workingDir}{Path to the working directory.}

\item{datasetDir}{Name of the directory containing the Dataset. The Directory containing the dataset has to be in the working directory 'workingDir'.}

\item{cores}{Number of cores to be used for parallel computation}
}
\value{
Creates a coexpression network and saves it tn the "datasetDir" directory.
}
\description{
Creates and stores a coexpression network from gene expression data based on WGCNA blockwiseModules.
}
\examples{
check README for an example
}
