% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexpressionNetwork.r
\name{CoexpressionNetwork_Clusteranalysis}
\alias{CoexpressionNetwork_Clusteranalysis}
\title{Clusteranalysis on the coexpression network}
\usage{
CoexpressionNetwork_Clusteranalysis(
  datasetDir,
  workingDir,
  networkType = "unsigned",
  power_thre
)
}
\arguments{
\item{datasetDir}{Name of the directory containing the dataset. The directory containing the dataset has to be in the working directory 'workingDir'.}

\item{workingDir}{Path to the working directory.}

\item{networkType}{Type of the previously computed coexpression network, default = unsigned.}

\item{power_thre}{Soft power threshold used for the coexpression network creation.}
}
\value{
Creates a .csv file containing the cluster membership for each gene, a .csv file containing all hub genes, a .csv file containing detailed information regarding the most common GO terms for each cluster, as well as a .csv file containing each genes correlation towards the specific phenotype.
}
\description{
This function identifies clusters (modules) in the coexpression network, as well as hub genes.
}
\examples{
check README for an example
}
