% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexpressionNetwork.r
\name{CoexpressionNetwork_Postprocessing}
\alias{CoexpressionNetwork_Postprocessing}
\title{Preparing Coexpression Network for NetRank.}
\usage{
CoexpressionNetwork_Postprocessing(
  datasetDir,
  workingDir,
  networkType = "unsigned",
  standardCorr_column = "PearsonCorrelation",
  geneName_col = "GeneName",
  dataFolder = "standardGeneScreeningResults",
  NetWorkCutThreshold,
  corrFilename
)
}
\arguments{
\item{datasetDir}{Name of the directory containing the dataset. The directory containing the dataset has to be in the working directory 'workingDir'.}

\item{workingDir}{Path to the working directory.}

\item{networkType}{Type of the previously computed coexpression network, default = "unsigned".}

\item{standardCorr_column}{Name of the column containing the correlation information in the correlation file. Default = 'PearsonCorrelation'.}

\item{geneName_col}{Name of the column containing the gene names in the correlation file. Default = 'GeneName'}

\item{dataFolder}{Name of the directory containing the correlation file. The NetRank results will be saved into this directory. Default = 'standardGeneScreeningResults'.}

\item{NetWorkCutThreshold}{In the coexpression network edges with a coexpression value below the threshold are discarded. This is necessary due to the size of the network, without the cut it is too large for further processing.}

\item{corrFilename}{Name of the file containing the gene correlation information.}
}
\value{
Saves an .RData file in the datasetDir directory for NetRank application.
}
\description{
Creates files for NetRank from the coexpression network
}
\examples{
check the README for an example
}
