% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexpressionNetwork.r
\name{CoexpressionNetwork_Preprocessing}
\alias{CoexpressionNetwork_Preprocessing}
\title{Outlier Detection for WGCNA}
\usage{
CoexpressionNetwork_Preprocessing(
  networkType = "unsigned",
  datasetDir,
  workingDir,
  pathToExprData,
  pathToTraitData,
  sampleIDCol = "sampleName",
  phenoCol = "pheno"
)
}
\arguments{
\item{networkType}{Type of coexpression network to be created. Default = "unsinged".}

\item{datasetDir}{Name of the directory containing the Dataset. The Directory containing the dataset has to be in the working directory 'workingDir'.}

\item{workingDir}{Path to the working directory.}

\item{pathToExprData}{Path to the csv file which contains the expression data.}

\item{pathToTraitData}{Path to the csv file which contains the phenotype data.}

\item{phenoCol}{Column name of phenotypes. Default = "Pheno".}

\item{sampleIdCol}{Column name of sample IDs. Default = "samplename".}
}
\value{
Creates and saves pdf files in workingDir/dataset/networkType/Plots/
}
\description{
Creates files to manually detect outliers for creating a the coexpression network. Plots are saved to workingDir/datasetDir/Plots.
}
\examples{
check README for an example
}
