% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinInteractionNetwork.R
\name{fetchStringDB}
\alias{fetchStringDB}
\title{Building a protein interaction network.}
\usage{
fetchStringDB(
  datasetDir,
  workingDir,
  standardCorr_column = "PearsonCorrelation",
  geneName_col = "GeneName",
  dataFolder = "standardGeneScreeningResults",
  corrFilename,
  species = 9606
)
}
\arguments{
\item{datasetDir}{Name of the directory containing the dataset. The directory containing the dataset has to be in the working directory 'workingDir'.}

\item{workingDir}{Path to the working directory.}

\item{standardCorr_column}{Name of the column containing the correlation information in the correlation file. Default = 'PearsonCorrelation'.}

\item{geneName_col}{Name of the column containing the gene names in the correlation file. Default = 'GeneName'}

\item{dataFolder}{Name of the directory containing the correlation File. The results will be saved into this directory. Default = 'standardGeneScreeningResults'.}

\item{corrFilename}{Name of the file containing the gene correlation information.}

\item{species}{STRINGdb ID of the species which the data is from. Default = 9606 for human.}
}
\value{
Saves files for NetRank application in the "datasetDir" directory.
}
\description{
Creates a protein interaction network from correlation data based on STRINGdb protein interaction information.
}
\examples{
check README for an example
}
