% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetRank.R
\name{RunNetRank}
\alias{RunNetRank}
\title{NetRank Algorithm}
\usage{
RunNetRank(
  network,
  datasetDir,
  workingDir,
  cores,
  d = 0.5,
  minError = 0.01,
  dataFolder = "standardGeneScreeningResults",
  NetWorkCutThreshold
)
}
\arguments{
\item{datasetDir}{Name of the directory containing the dataset. The directory containing the dataset has to be in the working directory 'workingDir'.}

\item{workingDir}{Path to the working directory.}

\item{cores}{Number of cores for parallelization.}

\item{d}{Dampening factor between 0 and 1. This factor weights the two parts of the NetRank formula against each other (node correlation vs. node connectivity). A d value close to 0 increases the importance of node correlation, a value close to 1 increases the importance of node connectivity. For further information check https://doi.org/10.3389/fbinf.2022.780229. Default = 0.5}

\item{minError}{Threshold for stopping the main Netrank computation as soon as the root mean square deviation of the netrank score between iterations is below minError. Default = 0.01}

\item{dataFolder}{Name of the directory containing the correlation File. The results will be saved into this directory. Default = 'standardGeneScreeningResults'.}

\item{NetWorkCutThreshold}{Previously chosen threshold for cutting edges with low coexpression values. Only relevant if network == "Coexpression".}

\item{networkType}{Choose between two network types "StringDB" or "Coexpression".}
}
\value{
Saves a file containing gene NetRank score and ranking, as well as the phenotype correlation. Results are saved in the 'dataFolder' in csv format.
}
\description{
Applies the NetRank algorithm to a previously constructed network.
}
\examples{
The Working directory has to contain a directory named 'datasetDir'. The 'datasetDir' directory has to contain a directory named 'dataFolder'.
}
