% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoexpressionNetwork.r
\name{SignatureClusters}
\alias{SignatureClusters}
\title{Fetches cluster information and hub genes from a gene signature}
\usage{
SignatureClusters(
  p_threshold,
  numberOfGenes,
  pathToSignatureFile,
  workingDir,
  datasetDir,
  rankCol = "rank_NetRank_score",
  networkType
)
}
\arguments{
\item{p_threshold}{p-value threshold for the gene signature. Genes with a p-value greater or equal to p_treshold are discarded.}

\item{numberOfGenes}{Number of genes that will be in the signature.}

\item{pathToSignatureFile}{Path to file that contains a ranked list of genes, i.e. the NetRank output file.}

\item{workingDir}{Path to the working directory.}

\item{datasetDir}{Name of the directory containing the dataset. The directory containing the dataset has to be in the working directory 'workingDir'.}

\item{networkType}{Type of the previously computed coexpression network, default = unsigned.}

\item{NetWorkCutThreshold}{Previously chosen threshold for cutting edges with low coexpression values.}
}
\value{
This function saves three files in a .csv format. The first file contains a signature of genes with full cluster information, the second file lists all hub genes found in the signature, the third file contains additional information about the clusters found in the signature.
}
\description{
This function takes a gene signature and outputs cluster information and hub genes for this signature.
}
\examples{
check the README for an example
}
